WIP: bracket reconciliation, game station, wagering, settlement, and client updates
Made-with: Cursor

+++ Service/Models/Requests/AdminUpdateGameOddsRequestModel.cs
@@ -2,10 +2,10 @@ namespace TecmoTourney.Models.Requests
 {
     public class AdminUpdateGameOddsRequestModel
     {
-        public int Spread { get; set; }
+        public decimal Spread { get; set; }
         public int? FavoredPlayerId { get; set; }
-        public int? MoneyLinePlayer1 { get; set; }
-        public int? MoneyLinePlayer2 { get; set; }
+        public decimal? MoneyLinePlayer1 { get; set; }
+        public decimal? MoneyLinePlayer2 { get; set; }
         public decimal? OverUnder { get; set; }
     }
 }
